+++ tests/disjoint_set.rs
@@ -1,13 +1,16 @@
 use disjoint::DisjointSet;
 
-fn verify_subsets(disjoint_set: &DisjointSet, expected_subsets: &[Vec<usize>]) {
+fn verify_subsets(disjoint_set: &DisjointSet, expected_subsets_ordered: &[Vec<usize>]) {
     assert_eq!(
         disjoint_set.len(),
-        expected_subsets.iter().map(|subset| subset.len()).sum()
+        expected_subsets_ordered
+            .iter()
+            .map(|subset| subset.len())
+            .sum()
     );
 
-    for (subset_id, subset) in expected_subsets.iter().enumerate() {
-        for (other_subset_id, other_subset) in expected_subsets.iter().enumerate() {
+    for (subset_id, subset) in expected_subsets_ordered.iter().enumerate() {
+        for (other_subset_id, other_subset) in expected_subsets_ordered.iter().enumerate() {
             for &member in subset {
                 for &other_member in other_subset {
                     assert_eq!(
@@ -18,11 +21,13 @@ fn verify_subsets(disjoint_set: &DisjointSet, expected_subsets: &[Vec<usize>]) {
             }
         }
     }
+
+    assert_eq!(disjoint_set.get_sets(), expected_subsets_ordered)
 }
 
 #[test]
 fn constructor_prodcues_singletons() {
-    let disjoint_set = DisjointSet::new(5);
+    let disjoint_set = DisjointSet::with_len(5);
 
     let expected_subsets = [vec![0], vec![1], vec![2], vec![3], vec![4]];
     verify_subsets(&disjoint_set, &expected_subsets);
@@ -31,7 +36,7 @@ fn constructor_prodcues_singletons() {
 #[test]
 fn constructor_constructs_correct_len_and_if_empty() {
     for size in 0..100 {
-        let disjoint_set = DisjointSet::new(size);
+        let disjoint_set = DisjointSet::with_len(size);
         assert_eq!(disjoint_set.len(), size);
         assert_eq!(disjoint_set.is_empty(), size == 0);
     }
@@ -40,7 +45,7 @@ fn constructor_constructs_correct_len_and_if_empty() {
 #[test]
 fn joining_and_is_joined_dont_change_len() {
     for size in 3..103 {
-        let mut disjoint_set = DisjointSet::new(size);
+        let mut disjoint_set = DisjointSet::with_len(size);
 
         disjoint_set.join(0, 1);
         assert!(!disjoint_set.is_joined(1, 2));
@@ -56,7 +61,7 @@ fn joining_and_is_joined_dont_change_len() {
 
 #[test]
 fn join_one_pair() {
-    let mut disjoint_set = DisjointSet::new(3);
+    let mut disjoint_set = DisjointSet::with_len(3);
     disjoint_set.join(0, 1);
     let expected_subsets = [vec![0, 1], vec![2]];
     verify_subsets(&disjoint_set, &expected_subsets);
@@ -64,7 +69,7 @@ fn join_one_pair() {
 
 #[test]
 fn join_element_with_itself() {
-    let mut disjoint_set = DisjointSet::new(3);
+    let mut disjoint_set = DisjointSet::with_len(3);
     disjoint_set.join(1, 1);
     let expected_subsets = [vec![0], vec![1], vec![2]];
     verify_subsets(&disjoint_set, &expected_subsets);
@@ -72,7 +77,7 @@ fn join_element_with_itself() {
 
 #[test]
 fn join_one_pair_twice() {
-    let mut disjoint_set = DisjointSet::new(4);
+    let mut disjoint_set = DisjointSet::with_len(4);
     disjoint_set.join(0, 1);
     disjoint_set.join(0, 1);
     let expected_subsets = [vec![0, 1], vec![2], vec![3]];
@@ -81,7 +86,7 @@ fn join_one_pair_twice() {
 
 #[test]
 fn join_one_pair_and_reversed() {
-    let mut disjoint_set = DisjointSet::new(2);
+    let mut disjoint_set = DisjointSet::with_len(2);
     disjoint_set.join(0, 1);
     disjoint_set.join(1, 0);
     let expected_subsets = [vec![0, 1]];
@@ -90,7 +95,7 @@ fn join_one_pair_and_reversed() {
 
 #[test]
 fn join_two_pairs_overlapping() {
-    let mut disjoint_set = DisjointSet::new(4);
+    let mut disjoint_set = DisjointSet::with_len(4);
     disjoint_set.join(0, 1);
     disjoint_set.join(1, 3);
     let expected_subsets = [vec![0, 1, 3], vec![2]];
@@ -99,13 +104,13 @@ fn join_two_pairs_overlapping() {
 
 #[test]
 fn join_two_pairs_non_overlapping() {
-    let mut disjoint_set = DisjointSet::new(10);
+    let mut disjoint_set = DisjointSet::with_len(10);
     disjoint_set.join(0, 4);
     disjoint_set.join(2, 7);
     let expected_subsets = [
         vec![0, 4],
-        vec![2, 7],
         vec![1],
+        vec![2, 7],
         vec![3],
         vec![5],
         vec![6],
@@ -117,7 +122,7 @@ fn join_two_pairs_non_overlapping() {
 
 #[test]
 fn join_two_non_trivial_subsets() {
-    let mut disjoint_set = DisjointSet::new(5);
+    let mut disjoint_set = DisjointSet::with_len(5);
     disjoint_set.join(0, 1);
     disjoint_set.join(2, 3);
     let expected_subsets = [vec![0, 1], vec![2, 3], vec![4]];
@@ -130,7 +135,7 @@ fn join_two_non_trivial_subsets() {
 
 #[test]
 fn arbitrary_sequence_of_join_and_is_joined() {
-    let mut disjoint_set = DisjointSet::new(5);
+    let mut disjoint_set = DisjointSet::with_len(5);
     assert!(!disjoint_set.is_joined(1, 2));
     assert!(!disjoint_set.is_joined(1, 2));
     assert!(disjoint_set.is_joined(1, 1));
@@ -175,49 +180,49 @@ fn arbitrary_sequence_of_join_and_is_joined() {
 #[test]
 #[should_panic]
 fn panic_join_first_element_oob() {
-    let mut disjoint_set = DisjointSet::new(1000);
+    let mut disjoint_set = DisjointSet::with_len(1000);
     disjoint_set.join(1000, 5);
 }
 
 #[test]
 #[should_panic]
 fn panic_join_second_element_oob() {
-    let mut disjoint_set = DisjointSet::new(100);
+    let mut disjoint_set = DisjointSet::with_len(100);
     disjoint_set.join(0, 1000000000000);
 }
 
 #[test]
 #[should_panic]
 fn panic_join_both_elements_oob() {
-    let mut disjoint_set = DisjointSet::new(0);
+    let mut disjoint_set = DisjointSet::new();
     disjoint_set.join(0, 0);
 }
 
 #[test]
 #[should_panic]
 fn panic_is_joined_first_element_oob() {
-    let disjoint_set = DisjointSet::new(1000);
+    let disjoint_set = DisjointSet::with_len(1000);
     let _ = disjoint_set.is_joined(1000, 5);
 }
 
 #[test]
 #[should_panic]
 fn panic_is_joined_second_element_oob() {
-    let disjoint_set = DisjointSet::new(100);
+    let disjoint_set = DisjointSet::with_len(100);
     let _ = disjoint_set.is_joined(0, 1000000000000);
 }
 
 #[test]
 #[should_panic]
 fn panic_is_joined_both_elements_oob() {
-    let disjoint_set = DisjointSet::new(0);
+    let disjoint_set = DisjointSet::new();
     let _ = disjoint_set.is_joined(0, 0);
 }
 
 #[test]
 #[should_panic]
 fn clone_clones() {
-    let mut disjoint_set = DisjointSet::new(5);
+    let mut disjoint_set = DisjointSet::with_len(5);
     disjoint_set.join(2, 4);
     disjoint_set.join(3, 4);
 
@@ -231,16 +236,16 @@ fn clone_clones() {
 
 #[test]
 fn different_len_not_equal() {
-    let l = DisjointSet::new(5);
-    let r = DisjointSet::new(6);
+    let l = DisjointSet::with_len(5);
+    let r = DisjointSet::with_len(6);
 
     assert_ne!(l, r);
 }
 
 #[test]
 fn same_len_not_equal() {
-    let mut l = DisjointSet::new(5);
-    let mut r = DisjointSet::new(5);
+    let mut l = DisjointSet::with_len(5);
+    let mut r = DisjointSet::with_len(5);
 
     l.join(2, 4);
     r.join(2, 3);
@@ -250,8 +255,8 @@ fn same_len_not_equal() {
 
 #[test]
 fn different_joining_order_equal() {
-    let mut l = DisjointSet::new(5);
-    let mut r = DisjointSet::new(5);
+    let mut l = DisjointSet::with_len(5);
+    let mut r = DisjointSet::with_len(5);
 
     l.join(2, 4);
     l.join(1, 3);
@@ -268,7 +273,7 @@ fn different_joining_order_equal() {
 
 #[test]
 fn get_sets_empty() {
-    let ds = DisjointSet::new(0);
+    let ds = DisjointSet::new();
 
     let sets = ds.get_sets();
 
@@ -279,7 +284,7 @@ fn get_sets_empty() {
 
 #[test]
 fn get_sets_singletons() {
-    let ds = DisjointSet::new(10);
+    let ds = DisjointSet::with_len(10);
 
     let sets = ds.get_sets();
 
@@ -301,7 +306,7 @@ fn get_sets_singletons() {
 
 #[test]
 fn get_sets_all_in_one() {
-    let mut ds = DisjointSet::new(8);
+    let mut ds = DisjointSet::with_len(8);
 
     ds.join(0, 1);
     ds.join(2, 3);
@@ -320,7 +325,7 @@ fn get_sets_all_in_one() {
 
 #[test]
 fn get_sets_complex() {
-    let mut ds = DisjointSet::new(5);
+    let mut ds = DisjointSet::with_len(5);
 
     ds.join(0, 3);
     ds.join(2, 1);
@@ -342,7 +347,7 @@ fn construct_with_capacity() {
 
 #[test]
 fn add_singleton_produces_singleton() {
-    let mut ds = DisjointSet::new(3);
+    let mut ds = DisjointSet::with_len(3);
     ds.join(0, 2);
     verify_subsets(&ds, &[vec![0, 2], vec![1]]);
     ds.add_singleton();
@@ -353,7 +358,7 @@ fn add_singleton_produces_singleton() {
 
 #[test]
 fn can_join_elements_added_later() {
-    let mut ds = DisjointSet::new(3);
+    let mut ds = DisjointSet::with_len(3);
 
     ds.join(0, 2);
 
@@ -372,4 +377,19 @@ fn can_join_elements_added_later() {
 #[should_panic]
 fn with_capacity_panics() {
     let _ = DisjointSet::with_capacity(isize::MAX as usize - 1);
+}
+
+#[test]
+fn different_ways_of_empty_construction() {
+    let empty_with_len = DisjointSet::with_len(0);
+    assert!(empty_with_len.is_empty());
+
+    let empty_new = DisjointSet::new();
+    assert!(empty_new.is_empty());
+
+    let empty_default = DisjointSet::default();
+    assert!(empty_new.is_empty());
+
+    assert_eq!(empty_with_len, empty_new);
+    assert_eq!(empty_with_len, empty_default);
 }